refactor(engine): decompose consumer.rs into submodules (SPL-25 #2453)
Split the 1306-line concurrent_delta/consumer.rs into focused submodules,
each well under the 650-line ceiling. Public surface unchanged: existing
re-exports continue to expose DeltaConsumer, DeltaConsumerStats, and the
two iterator types from concurrent_delta::consumer.

Layout:
- consumer/mod.rs     - public types, impl block, iterators (332 lines)
- consumer/spawn.rs   - ReorderMode selector and spawn_inner plumbing (176)
- consumer/loops.rs   - bare and spillable reorder loops, drain helpers (239)
- consumer/tests.rs   - integration tests moved verbatim (617)
